requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0

# Database & Async
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22

# Google OAuth
authlib==1.3.0
httpx==0.25.2

# Redis
redis==7.1.1
aioredis==2.0.1

# Environment
python-dotenv==1.2.1
